[feat] MaPageUI 구현 완료
