+++ YU-LikeLion-iOS/View/MyPage/ImagePickerView.swift
@@ -1,0 +1,63 @@
+//
+//  ImagePickerView.swift
+//  YU-LikeLion-iOS
+//
+//  Created by 윤강록 on 2/26/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct UIImagePicker: UIViewControllerRepresentable {
+    
+    typealias UIViewControllerType = UIImagePickerController
+    
+    @Environment(\.presentationMode)
+    private var presentationMode // 해당 뷰컨트롤러의 노출 여부
+    let sourceType: UIImagePickerController.SourceType
+    let imagePicked: (UIImage) -> () // 이미지가 선택됐을때 결과 호출
+    
+
+    
+    class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
+        
+        let parent: UIImagePicker
+        
+        init(parent: UIImagePicker) {
+            self.parent = parent
+        }
+        
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+            
+            if let image = info[.originalImage] as? UIImage {
+                parent.imagePicked(image)
+                parent.presentationMode.wrappedValue.dismiss()
+            }
+            
+        }
+        
+        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+            parent.presentationMode.wrappedValue.dismiss()
+        }
+    }
+    
+    func makeCoordinator() -> Coordinator {
+        Coordinator(parent: self)
+    }
+    
+    
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        
+        let picker = UIImagePickerController()
+        
+        picker.delegate = context.coordinator
+        
+        return picker
+        
+    }
+    
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {
+        
+    }
+
+}
