+++ YU-LikeLion-iOS/View/MyPage/ModifyMyPageView.swift
@@ -8,10 +8,37 @@
 import SwiftUI
 
 struct ModifyMyPageView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        @State var showingPhoto = false
+        @State var image : UIImage?
+        
+        var body: some View {
+            VStack {
+                ZStack {
+                    Rectangle()
+                        .fill(Color.main)
+                        .frame(width: 398, height: 300)
+                    
+                    Button(action: {
+                        self.showingPhoto.toggle()
+                    }, label: {
+                        let selected = image == nil ? Image("Image") : Image(uiImage : image!)
+                        selected
+                            .resizable()
+                            .frame(width: 150, height: 150)
+                            
+                    })
+                    .clipShape(Circle())
+                    .sheet(isPresented: $showingPhoto) {
+                        UIImagePicker(sourceType: .photoLibrary) { (image) in
+                            self.image = image
+                        }
+                    }
+                    
+                }
+             
+            }
+        }
     }
-}
 
 #Preview {
     ModifyMyPageView()
